requirements.txt
PyQt5==5.15.10
Pillow==10.1.0
reportlab==4.0.7
PyMuPDF==1.23.8
